Remove Actions from DependencyCarrier
Register `ACTION`s in `ActionModule` and  use `NodeClient.execute()`
in the respective plans.
