+++ server/src/main/java/io/crate/execution/TransportExecutorModule.java
@@ -23,13 +23,8 @@ package io.crate.execution;
 
 import org.elasticsearch.common.inject.AbstractModule;
 
-import io.crate.execution.ddl.TransportSwapRelationsAction;
 import io.crate.execution.ddl.index.TransportSwapAndDropIndexNameAction;
-import io.crate.execution.ddl.views.TransportCreateViewAction;
-import io.crate.execution.ddl.views.TransportDropViewAction;
 import io.crate.execution.jobs.JobSetup;
-import io.crate.expression.udf.TransportCreateUserDefinedFunctionAction;
-import io.crate.expression.udf.TransportDropUserDefinedFunctionAction;
 import io.crate.lucene.LuceneQueryBuilder;
 import io.crate.replication.logical.action.TransportAlterPublicationAction;
 import io.crate.replication.logical.action.TransportCreatePublicationAction;
@@ -45,11 +40,6 @@ public class TransportExecutorModule extends AbstractModule {
         bind(LuceneQueryBuilder.class).asEagerSingleton();
 
         bind(TransportSwapAndDropIndexNameAction.class).asEagerSingleton();
-        bind(TransportCreateUserDefinedFunctionAction.class).asEagerSingleton();
-        bind(TransportDropUserDefinedFunctionAction.class).asEagerSingleton();
-        bind(TransportCreateViewAction.class).asEagerSingleton();
-        bind(TransportDropViewAction.class).asEagerSingleton();
-        bind(TransportSwapRelationsAction.class).asEagerSingleton();
         bind(TransportAnalyzeAction.class).asEagerSingleton();
 
         bind(TransportCreatePublicationAction.class).asEagerSingleton();
