+++ server/src/main/java/io/crate/expression/udf/TransportCreateUserDefinedFunction.java
@@ -21,7 +21,10 @@
 
 package io.crate.expression.udf;
 
+import java.io.IOException;
+
 import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.ActionType;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.action.support.master.TransportMasterNodeAction;
 import org.elasticsearch.cluster.ClusterState;
@@ -29,29 +32,37 @@ import org.elasticsearch.cluster.block.ClusterBlockException;
 import org.elasticsearch.cluster.block.ClusterBlockLevel;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.inject.Inject;
-import org.elasticsearch.common.inject.Singleton;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
 
-import java.io.IOException;
+import io.crate.exceptions.ScriptException;
 
-@Singleton
-public class TransportDropUserDefinedFunctionAction
-    extends TransportMasterNodeAction<DropUserDefinedFunctionRequest, AcknowledgedResponse> {
+public class TransportCreateUserDefinedFunction
+    extends TransportMasterNodeAction<CreateUserDefinedFunctionRequest, AcknowledgedResponse> {
 
+    public static final Action ACTION = new Action();
     private final UserDefinedFunctionService udfService;
 
+    public static class Action extends ActionType<AcknowledgedResponse> {
+        private static final String NAME = "internal:crate:sql/udf/create";
+
+        private Action() {
+            super(NAME);
+        }
+    }
+
     @Inject
-    public TransportDropUserDefinedFunctionAction(TransportService transportService,
-                                                  ClusterService clusterService,
-                                                  ThreadPool threadPool,
-                                                  UserDefinedFunctionService udfService) {
-        super("internal:crate:sql/udf/drop",
+    public TransportCreateUserDefinedFunction(TransportService transportService,
+                                              ClusterService clusterService,
+                                              ThreadPool threadPool,
+                                              UserDefinedFunctionService udfService) {
+        super(
+            ACTION.name(),
             transportService,
             clusterService,
             threadPool,
-            DropUserDefinedFunctionRequest::new
+            CreateUserDefinedFunctionRequest::new
         );
         this.udfService = udfService;
     }
@@ -67,21 +78,19 @@ public class TransportDropUserDefinedFunctionAction
     }
 
     @Override
-    protected void masterOperation(final DropUserDefinedFunctionRequest request,
+    protected void masterOperation(final CreateUserDefinedFunctionRequest request,
                                    ClusterState state,
                                    ActionListener<AcknowledgedResponse> listener) throws Exception {
-        udfService.dropFunction(
-            request.schema(),
-            request.name(),
-            request.argumentTypes(),
-            request.ifExists(),
-            listener,
-            request.masterNodeTimeout()
-        );
+        UserDefinedFunctionMetadata metadata = request.userDefinedFunctionMetadata();
+        String errorMessage = udfService.getLanguage(metadata.language()).validate(metadata);
+        if (errorMessage != null) {
+            throw new ScriptException(errorMessage, metadata.language());
+        }
+        udfService.registerFunction(metadata, request.replace(), listener, request.masterNodeTimeout());
     }
 
     @Override
-    protected ClusterBlockException checkBlock(DropUserDefinedFunctionRequest request, ClusterState state) {
+    protected ClusterBlockException checkBlock(CreateUserDefinedFunctionRequest request, ClusterState state) {
         return state.blocks().globalBlockedException(ClusterBlockLevel.METADATA_READ);
     }
 }
